+++ lib/src/requests/delete_all_paragraph_tab_stops_online_request.dart
@@ -100,7 +100,7 @@ class DeleteAllParagraphTabStopsOnlineRequest implements RequestBase {
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
     var _result = DeleteAllParagraphTabStopsOnlineResponse();
-    _result.deserialize(_apiClient.deserializeMultipartMap(_body));
+    _result.deserialize(_apiClient, _apiClient.deserializeMultipartMap(_body));
     return _result;
   }
 }