+++ lib/src/requests/delete_comment_online_request.dart
@@ -109,6 +109,5 @@ class DeleteCommentOnlineRequest implements RequestBase {
 
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
-    return _body;
   }
 }