+++ lib/src/requests/delete_footnote_online_request.dart
@@ -113,6 +113,5 @@ class DeleteFootnoteOnlineRequest implements RequestBase {
 
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
-    return _body;
   }
 }