+++ lib/src/requests/delete_run_online_request.dart
@@ -117,6 +117,5 @@ class DeleteRunOnlineRequest implements RequestBase {
 
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
-    return _body;
   }
 }