+++ lib/src/requests/delete_table_cell_online_request.dart
@@ -117,6 +117,5 @@ class DeleteTableCellOnlineRequest implements RequestBase {
 
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
-    return _body;
   }
 }