+++ lib/src/requests/insert_watermark_image_online_request.dart
@@ -127,7 +127,7 @@ class InsertWatermarkImageOnlineRequest implements RequestBase {
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
     var _result = InsertWatermarkImageOnlineResponse();
-    _result.deserialize(_apiClient.deserializeMultipartMap(_body));
+    _result.deserialize(_apiClient, _apiClient.deserializeMultipartMap(_body));
     return _result;
   }
 }