+++ lib/src/requests/update_drawing_object_online_request.dart
@@ -134,7 +134,7 @@ class UpdateDrawingObjectOnlineRequest implements RequestBase {
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
     var _result = UpdateDrawingObjectOnlineResponse();
-    _result.deserialize(_apiClient.deserializeMultipartMap(_body));
+    _result.deserialize(_apiClient, _apiClient.deserializeMultipartMap(_body));
     return _result;
   }
 }