+++ lib/src/responses/update_table_cell_format_online_response.dart
@@ -31,6 +31,8 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import '../../aspose_words_cloud.dart';
+import '../api_client.dart';
+import '../body_part_data.dart';
 
 /// Updates the formatting properties of a cell in the table row.
 class UpdateTableCellFormatOnlineResponse implements ResponseBase {
@@ -38,16 +40,16 @@ class UpdateTableCellFormatOnlineResponse implements ResponseBase {
   TableCellFormatResponse model;
 
   /// The document after modification.
-  ByteData document;
+  Map<String, ByteData> document;
 
   @override
-  void deserialize(Map<String, ByteData> _parts) {
+  void deserialize(ApiClient apiClient, Map<String, BodyPartData> _parts) {
     model = TableCellFormatResponse();
     final _modelBody = _parts['model'];
-    final _modelJsonData = utf8.decode(_modelBody.buffer.asUint8List(_modelBody.offsetInBytes, _modelBody.lengthInBytes));
+    final _modelJsonData = utf8.decode(_modelBody.content.buffer.asUint8List(_modelBody.content.offsetInBytes, _modelBody.content.lengthInBytes));
     model.deserialize(jsonDecode(_modelJsonData) as Map<String, dynamic>);
 
-    document = _parts['document'];
+    document = apiClient.deserializeFilesCollection(_parts['document']);
   }
 }